LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-size-animation-expected.txt


FAIL Animation with anchor-size() assert_equals: expected 150 but got 50
